quick-and-dirty change to retry setting a link when it failed
The failure occurs when the parent (Epic link) is not yet created when the first revision of a 'child' is saved. The key is to _not_ configure the _child link type_, because that might result in attaching the link to another Epic (if it was attached to one somewhere in the history).